Replication Test

+++ src/main/java/ProjectY/Client/Client.java
@@ -167,9 +167,15 @@ public class Client {
 
     public void askReplicationFiles(String newNode, String newNodeIP) {
         for(FileLog fileLog : fileLogList){
-            if (currentID < Hash(newNode) && Hash(newNode) < fileLog.getFileID()) {
-                tcpModule.sendFile(fileLog.getOwnerIP(), newNodeIP, fileLog.getFileName());
-                deleteFile(fileLog.getFileName());
+            if(fileLog.getOwnerIP()==IPAddres){
+                if(Hash(newNode) < fileLog.getFileID()){
+                    tcpModule.sendFile(fileLog.getOwnerIP(), newNodeIP, fileLog.getFileName());
+                }
+            }else {
+                if (currentID < Hash(newNode) && Hash(newNode) < fileLog.getFileID()) {
+                    tcpModule.sendFile(fileLog.getOwnerIP(), newNodeIP, fileLog.getFileName());
+                    deleteFile(fileLog.getFileName());
+                }
             }
 
         }
